+++ R/1.R
@@ -1,12 +1,12 @@
 # this is stuff that needs to be run first and in order for S4 crap to work
 methods::setOldClass("telemetry")
-new.telemetry <- methods::setClass("telemetry", representation(info="list"), contains="data.frame")
-
-methods::setOldClass("UD")
-new.UD <- methods::setClass("UD", representation(info="list"), contains="list")
+new.telemetry <- methods::setClass("telemetry", representation(info="list",UERE="list"), contains="data.frame")
 
 methods::setOldClass("ctmm")
 new.ctmm <- methods::setClass("ctmm", representation(info="list"), contains="list")
+
+methods::setOldClass("UD")
+new.UD <- methods::setClass("UD", representation(info="list",type="character",CTMM="ctmm"), contains="list")
 
 methods::setOldClass("variogram")
 new.variogram <- methods::setClass("variogram",representation("data.frame",info="list"))