+++ R/krige.R
@@ -306,6 +306,7 @@ fill.data <- function(data,CTMM=ctmm(tau=Inf),verbose=FALSE,t=NULL,dt=NULL,res=1
 #################################
 occurrence <- function(data,CTMM,H=0,res.time=10,res.space=10,grid=NULL,cor.min=0.5,dt.max=NULL)
 {
+  CTMM0 <- CTMM
   dt <- stats::median(diff(data$t))
 
   if(length(H)==1) { H <- diag(H,2) }
@@ -375,7 +376,7 @@ occurrence <- function(data,CTMM,H=0,res.time=10,res.space=10,grid=NULL,cor.min=
   # using the same data format as AKDE, but with only the ML estimate (alpha=1)
   KDE <- kde(data,H=H,W=w.grid,dr=dr/res.space)
   KDE$H <- diag(0,2)
-  KDE <- new.UD(KDE,info=info)
+  KDE <- new.UD(KDE,info=info,type='occurrence',CTMM=CTMM0)
   return(KDE)
 }